+++ muna/beta/openai/audio.py
@@ -6,12 +6,14 @@
 from ...services import PredictorService, PredictionService
 from ..remote.remote import RemotePredictionService
 from .speech import SpeechService
+from .transcription import TranscriptionService
 
 class AudioService:
     """
     Audio service.
     """
     speech: SpeechService
+    transcriptions: TranscriptionService
 
     def __init__(
         self,
@@ -20,3 +22,4 @@ class AudioService:
         remote_predictions: RemotePredictionService
     ):
         self.speech = SpeechService(predictors, predictions, remote_predictions)
+        self.transcriptions = TranscriptionService(predictors, predictions, remote_predictions)